# src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: "Api Module contract"
  version: 1.0.1
servers:
  - url: /api-api/v1
security:
  - basicAuth: []
tags:
  - name: Events
paths:
  /events:
    get:
      tags:
        - Events
      operationId: getEvents
      summary: Returns the information from all the events in a list
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/token'
      responses:
        200:
          $ref: '#/components/responses/EventsResponse'
        500:
          $ref: '#/components/responses/InternalServerErrorResponse'
  /event/{event-id}:
    get:
      tags:
        - Events
      operationId: getEvent
      summary: Returns the information from a specific event
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/token'
        - name: event-id
          description: event id
          in: path
          required: true
          schema:
            type: number
      responses:
        200:
          $ref: '#/components/responses/EventDetailsResponse'
        500:
          $ref: '#/components/responses/InternalServerErrorResponse'
  /pay:
    post:
      tags:
        - Events
      operationId: pay
      summary: Initiates the pay process
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/token'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayRequest'
      responses:
        200:
          $ref: '#/components/responses/PayResponse'
        500:
          $ref: '#/components/responses/InternalServerErrorResponse'
    x-codegen-request-body-name: body
components:
  parameters:
    userId:
      name: X-Otp-Auth-User-Id
      in: header
      required: true
      schema:
        type: string
    token:
      name: X-Otp-Auth-Token
      in: header
      required: true
      schema:
        type: string
  responses:
    PayResponse:
      description: Pay Response
      content:
        application/json:
          schema:
            title: Reservation Response
            type: object
            properties:
              reservationId:
                type: string
              info:
                $ref: '#/components/schemas/Info'
    EventDetailsResponse:
      description: Event Details Response
      content:
        application/json:
          schema:
            title: Events Details Response
            type: object
            properties:
              seats:
                type: array
                items:
                  $ref: '#/components/schemas/Seat'
              info:
                $ref: '#/components/schemas/Info'
    EventsResponse:
      description: Event Query Response
      content:
        application/json:
          schema:
            title: Events Response
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
              info:
                $ref: '#/components/schemas/Info'
    InternalServerErrorResponse:
      description: Internal Server Error Response
      content:
        application/json:
          schema:
            title:  Internal Server Error Response
            type: object
            properties:
              error:
                $ref: '#/components/schemas/Info'
  schemas:
    Info:
      required:
        - success
        - state
      type: object
      properties:
        success:
          type: boolean
        state:
          $ref: '#/components/schemas/State'
    State:
      type: object
      properties:
        code:
          type: integer
          format: int64
        reason:
          type: string
    PayRequest:
      required:
        - eventId
        - seatId
        - cardId
      type: object
      properties:
        eventId:
          type: number
        seatId:
          type: string
        cardId:
          type: string
    Seat:
      type: object
      properties:
        id:
          type: string
        currency:
          type: string
        price:
          type: number
        reserved:
          type: boolean
    Event:
      type: object
      properties:
        eventId:
          type: number
        title:
          type: string
        location:
          type: string
        startTimeStamp:
          type: string
          format: date-time
        endTimeStamp:
          type: string
          format: date-time
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic